+++ app/src/main/java/org/simple/clinic/overdue/PotentialAppointmentDate.kt
@@ -1,0 +1,12 @@
+package org.simple.clinic.overdue
+
+import org.threeten.bp.LocalDate
+
+data class PotentialAppointmentDate(
+    val scheduledFor: LocalDate,
+    val timeToAppointment: TimeToAppointment
+) : Comparable<PotentialAppointmentDate> {
+  override fun compareTo(other: PotentialAppointmentDate): Int {
+    return this.scheduledFor.compareTo(other.scheduledFor)
+  }
+}
